Adding a BcSMIMECryptographer
This is a first push.  Some clean up included.

I am testing this on a Windows 11 machine and the expected PKCS#1 v1.5 padding was expected with the original SMIMECryptographer.  But it is not.  There is a suspicion that new Windows fixes the NIST 800-131A rev 2 recomendation.  I am going to now test this on an older Windows version.